refactor: move mlx_dit and mlx_vae into models/mlx
- Relocate acestep/mlx_dit/ -> acestep/models/mlx/dit_*.py
- Relocate acestep/mlx_vae/ -> acestep/models/mlx/vae_*.py
- Consolidate duplicate mlx_available() into models/mlx/__init__.py
- Update all import paths in handler.py, diffusion.py, generation.py
- Delete old mlx_dit/ and mlx_vae/ directories
- No logic changes; pure move-and-rename refactor
